requirements.txt
beautifulsoup4==4.6.0
billiard==3.5.0.3
certifi==2017.11.5
chardet==3.0.4
Django==2.0
django-cors-headers==2.1.0
django-environ==0.4.4
django-filter==1.1.0
djangorestframework==3.7.3
ephem==3.7.6.0
idna==2.6
model-mommy==1.4.0
psycopg2==2.7.3.2
pytz==2017.3
requests==2.18.4
serpy==0.3.1
six==1.11.0
urllib3==1.22
vine==1.1.4
